Remove use of context, port to new rrq
Not yet in the mrc-ide drat because it will break on the cluster,
this includes manual installatin of rrq (though this is already
being done for other packages). Beacuse of the way we were using
context this is a pretty lightweight change to the queue creation.

Note that this change does not support the ability to spawn workers
externally